+++ lib/src/youtube_player.dart
@@ -1,6 +1,8 @@
 import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:youtube_player_flutter/src/fullscreen_youtube_player.dart';
@@ -29,6 +31,15 @@ enum PlayerState {
   PAUSED,
   BUFFERING,
   CUED,
+}
+
+/// Playback Rate or Speed for the video.
+enum PlaybackRate {
+  QUARTER,
+  HALF,
+  NORMAL,
+  ONE_AND_A_HALF,
+  DOUBLE,
 }
 
 typedef YoutubePlayerControllerCallback(YoutubePlayerController controller);
@@ -163,7 +174,9 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
 
   String _currentVideoId;
 
-  var _inFullScreen = false;
+  bool _inFullScreen = false;
+
+  bool _firstLoad = true;
 
   @override
   void initState() {
@@ -201,6 +214,17 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
   }
 
   void listener() async {
+    if (controller.value.isReady &&
+        controller.value.isEvaluationReady &&
+        _firstLoad) {
+      _firstLoad = false;
+      widget.flags.autoPlay
+          ? controller.load(startAt: widget.startAt.inSeconds)
+          : controller.cue(startAt: widget.startAt.inSeconds);
+      if (widget.flags.mute) {
+        controller.mute();
+      }
+    }
     if (controller.value.isFullScreen && !_inFullScreen) {
       _inFullScreen = true;
       Duration pos = await showFullScreenYoutubePlayer(
@@ -261,6 +285,45 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
     );
   }
 
+  Widget _thumbWidget() {
+    return CachedNetworkImage(
+      imageUrl: widget.thumbnailUrl ??
+          YoutubePlayer.getThumbnail(
+            videoId: controller.initialSource,
+          ),
+      fit: BoxFit.cover,
+      errorWidget: (context, url, _) {
+        return Container(
+          color: Colors.black,
+          child: Center(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.spaceAround,
+              children: [
+                Text(
+                  'Oops! Something went wrong!',
+                  style: TextStyle(
+                    fontWeight: FontWeight.w400,
+                    color: Colors.white,
+                  ),
+                ),
+                Text(
+                  'Might be an internet issue',
+                  style: TextStyle(
+                    fontWeight: FontWeight.w300,
+                    color: Colors.white,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        );
+      },
+      placeholder: (context, _) => Container(
+        color: Colors.black,
+      ),
+    );
+  }
+
   Widget _buildPlayer(double _aspectRatio) {
     return AspectRatio(
       aspectRatio: _aspectRatio,
@@ -271,7 +334,6 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
           _Player(
             controller: controller,
             flags: widget.flags,
-            startAt: widget.startAt,
           ),
           if (!controller.value.hasPlayed &&
               controller.value.playerState == PlayerState.BUFFERING)
@@ -279,13 +341,7 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
               color: Colors.black,
             ),
           if (!controller.value.hasPlayed && !widget.flags.hideThumbnail)
-            Image.network(
-              widget.thumbnailUrl ??
-                  YoutubePlayer.getThumbnail(
-                    videoId: controller.initialSource,
-                  ),
-              fit: BoxFit.cover,
-            ),
+            _thumbWidget(),
           if (!widget.flags.hideControls)
             TouchShutter(
               controller,
@@ -373,7 +429,8 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
 /// [ValueNotifier] for [YoutubePlayerController].
 class YoutubePlayerValue {
   YoutubePlayerValue({
-    @required this.isReady,
+    this.isReady = false,
+    this.isEvaluationReady = false,
     this.isLoaded = false,
     this.hasPlayed = false,
     this.duration = const Duration(),
@@ -383,12 +440,16 @@ class YoutubePlayerValue {
     this.isFullScreen = false,
     this.volume = 100,
     this.playerState = PlayerState.UNKNOWN,
+    this.playbackRate = PlaybackRate.NORMAL,
     this.errorCode = 0,
     this.webViewController,
   });
 
   /// This is true when underlying web player reports ready.
   final bool isReady;
+
+  /// This is true when JavaScript evaluation can be triggered.
+  final bool isEvaluationReady;
 
   /// This is true once video loads.
   final bool isLoaded;
@@ -417,6 +478,9 @@ class YoutubePlayerValue {
   /// The current state of the player defined as [PlayerState].
   final PlayerState playerState;
 
+  /// The current video playback rate defined as [PlaybackRate].
+  final PlaybackRate playbackRate;
+
   /// Reports the error code as described [here](https://developers.google.com/youtube/iframe_api_reference#Events).
   /// See the onError Section.
   final int errorCode;
@@ -429,6 +493,7 @@ class YoutubePlayerValue {
 
   YoutubePlayerValue copyWith({
     bool isReady,
+    bool isEvaluationReady,
     bool isLoaded,
     bool hasPlayed,
     Duration duration,
@@ -438,11 +503,13 @@ class YoutubePlayerValue {
     bool isFullScreen,
     double volume,
     PlayerState playerState,
+    PlaybackRate playbackRate,
     int errorCode,
     WebViewController webViewController,
   }) {
     return YoutubePlayerValue(
       isReady: isReady ?? this.isReady,
+      isEvaluationReady: isEvaluationReady ?? this.isEvaluationReady,
       isLoaded: isLoaded ?? this.isLoaded,
       duration: duration ?? this.duration,
       hasPlayed: hasPlayed ?? this.hasPlayed,
@@ -452,6 +519,7 @@ class YoutubePlayerValue {
       isFullScreen: isFullScreen ?? this.isFullScreen,
       volume: volume ?? this.volume,
       playerState: playerState ?? this.playerState,
+      playbackRate: playbackRate ?? this.playbackRate,
       errorCode: errorCode ?? this.errorCode,
       webViewController: webViewController ?? this.webViewController,
     );
@@ -461,6 +529,7 @@ class YoutubePlayerValue {
   String toString() {
     return '$runtimeType('
         'isReady: $isReady, '
+        'isEvaluationReady: $isEvaluationReady, '
         'isLoaded: $isLoaded, '
         'duration: $duration, '
         'position: $position, '
@@ -468,6 +537,7 @@ class YoutubePlayerValue {
         'isPlaying: $isPlaying, '
         'volume: $volume, '
         'playerState: $playerState, '
+        'playbackRate: $playbackRate, '
         'errorCode: $errorCode)';
   }
 }
@@ -484,7 +554,12 @@ class YoutubePlayerController extends ValueNotifier<YoutubePlayerValue> {
   }
 
   /// Hide YouTube Player annotations like title, share button and youtube logo.
-  void forceHideAnnotation() => _evaluateJS('hideAnnotations()');
+  /// It's hidden by default for iOS.
+  void forceHideAnnotation() {
+    if (Platform.isAndroid) {
+      _evaluateJS('hideAnnotations()');
+    }
+  }
 
   /// Plays the video.
   void play() => _evaluateJS('play()');
@@ -522,43 +597,31 @@ class YoutubePlayerController extends ValueNotifier<YoutubePlayerValue> {
     value = value.copyWith(position: position);
   }
 
-  /// Forces to enter fullScreen.
-  void enterFullScreen() {
-    value = value.copyWith(isFullScreen: true);
-    /*pause();
-    value = value.copyWith(isFullScreen: true);
-    Future.delayed(Duration(milliseconds: 500), () {
-      SystemChrome.setPreferredOrientations([
-        DeviceOrientation.landscapeRight,
-        DeviceOrientation.landscapeLeft,
-        if (autoRotationEnabled) ...[
-          DeviceOrientation.portraitUp,
-          DeviceOrientation.portraitDown,
-        ]
-      ]);
-      SystemChrome.setEnabledSystemUIOverlays([]);
-    });
-    Future.delayed(Duration(seconds: 2), () {
-      play();
-    });*/
+  /// Sets the size in pixels of the player.
+  void setSize(Size size) =>
+      _evaluateJS('setSize(${size.width * 100},${size.height * 100})');
+
+  void setPlaybackRate(PlaybackRate rate) {
+    switch (rate) {
+      case PlaybackRate.QUARTER:
+        _evaluateJS('setPlaybackRate(0.25)');
+        break;
+      case PlaybackRate.HALF:
+        _evaluateJS('setPlaybackRate(0.5)');
+        break;
+      case PlaybackRate.NORMAL:
+        _evaluateJS('setPlaybackRate(1)');
+        break;
+      case PlaybackRate.ONE_AND_A_HALF:
+        _evaluateJS('setPlaybackRate(1.5)');
+        break;
+      case PlaybackRate.DOUBLE:
+        _evaluateJS('setPlaybackRate(2)');
+        break;
+    }
   }
 
-  /// Forces to exit fullScreen.
-  void exitFullScreen() {
-    value = value.copyWith(isFullScreen: false);
-    /*pause();
-    value = value.copyWith(isFullScreen: false);
-    Future.delayed(Duration(milliseconds: 500), () {
-      SystemChrome.setPreferredOrientations([
-        DeviceOrientation.portraitUp,
-        DeviceOrientation.portraitDown,
-        DeviceOrientation.landscapeRight,
-        DeviceOrientation.landscapeLeft,
-      ]);
-      SystemChrome.setEnabledSystemUIOverlays(SystemUiOverlay.values);
-    });
-    Future.delayed(Duration(seconds: 2), () {
-      play();
-    });*/
-  }
+  void enterFullScreen() => value = value.copyWith(isFullScreen: true);
+
+  void exitFullScreen() => value = value.copyWith(isFullScreen: false);
 }