+++ lib/src/youtube_player_flags.dart
@@ -33,6 +33,8 @@ class YoutubePlayerFlags {
   ///
   /// Forcing annotation to hide is a hacky way. Although this shouldn't be against Youtube TOS, the author doesn't guarantee
   /// and won't be responsible for any casualties regarding the YouTube TOS violation.
+  ///
+  /// It's hidden by default on iOS. Changing this flag will have no effect on iOS.
   final bool forceHideAnnotation;
 
   /// Hides thumbnail if true