+++ CHANGELOG.md
@@ -1,5 +1,19 @@
-## 3.0.0+1
-* Orientation Issue Fixed.
+## 4.1.0
+* **Feature Added** Button to change playback rate in player.
+* Added `setPlaybackRate` method to `YoutubePlayerController`. [Issue #48](https://github.com/sarbagyastha/youtube_player_flutter/issues/48) 
+* **Improvement** Playback behaviour synced with `AppLifeCycle` [Issue #41](https://github.com/sarbagyastha/youtube_player_flutter/issues/41) 
+
+## 4.0.0+2
+* Finally working on iOS.
+* Fix error reports in LiveUI. [PR#44](https://github.com/sarbagyastha/youtube_player_flutter/pull/44)
+* Added `startAt` property.
+
+## 4.0.0+1
+**BREAKING CHANGES** 
+* `YoutubePLayerScaffold` has been removed.
+* Improved Orientations
+* Uses `webview_flutter` instead of `ytview`
+
 
 ## 3.0.0
 * **BREAKING CHANGES** See Migrating to 3.x.x for detail.